ITSAAS-43 - Автоматическая генерация сертификатов

+++ lms/djangoapps/certificates/api.py
@@ -1202,9 +1202,9 @@ def create_pdf_for_certificate(cert):
     tmp_file = str(uuid4()) + '.pdf'
     tmp_pdf_path = os.path.join(tempfile.gettempdir(), tmp_file)
 
-    subprocess.call(settings.PDF_RENDER_BIN + ' ' + tmp_html_path + ' ' + tmp_pdf_path, shell=True)
+    subprocess.call(settings.PDF_RENDER_BIN + ' -B 0 -L 0 -R 0 -T 0 ' + tmp_html_path + ' ' + tmp_pdf_path, shell=True)
 
-    storage_file_hash = str(cert.user.id) + '_' + hashlib.md5(str(cert.course_id)).hexdigest()
+    storage_file_hash = hashlib.md5(str(cert.course_id) + '/' + str(cert.user.id)).hexdigest()
     storage_class = settings.DEFAULT_FILE_STORAGE
     kwargs = {}
     if storage_class in ('storages.backends.s3boto.S3BotoStorage', 'openedx.core.storage.S3ReportStorage'):
@@ -1214,22 +1214,23 @@ def create_pdf_for_certificate(cert):
             'custom_domain': settings.AWS_S3_CUSTOM_DOMAIN
         }
     storage = get_storage(storage_class, **kwargs)
-    storage_certificate_file_path = storage_file_hash + '.pdf'
+    storage_certificate_file_path = str(cert.user.id) + '_' + storage_file_hash + '.pdf'
 
     if storage.exists(storage_certificate_file_path):
         storage.delete(storage_certificate_file_path)
 
-    with open(tmp_pdf_path, 'r') as pdf_file:
-        res = storage.save(storage_certificate_file_path, pdf_file)
+    if os.path.isfile(tmp_pdf_path):
+        with open(tmp_pdf_path, 'r') as pdf_file:
+            storage.save(storage_certificate_file_path, pdf_file)
 
-    url_to_save = storage.url(storage_certificate_file_path)
+        url_to_save = storage.url(storage_certificate_file_path)
 
-    cert.download_url = url_to_save
-    cert.download_uuid = storage_file_hash
-    cert.save()
+        cert.download_url = url_to_save
+        cert.download_uuid = storage_file_hash
+        cert.save()
 
+        os.remove(tmp_pdf_path)
     os.remove(tmp_html_path)
-    os.remove(tmp_pdf_path)
 
 
 def render_html(request, user_id, course_id):
